Added application files
The Info.plist is used by the osx-app-bundle maven plugin as a
template, this allows us to add a splash screen. The splash screen is
also used in the assembled jar and thus windows exe. The
mnb-application pom was modified to include the Info.plist and the new
splash image in the jar.